[move] #82 recordinfo
