+++ data/video/src/main/java/com/record/video/model/remote/response/RecordInfo.kt
@@ -1,4 +1,4 @@
-package com.record.network.model
+package com.record.video.model.remote.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
