+++ data/video/src/main/java/com/record/video/model/remote/response/ResponseGetVideoDto.kt
@@ -1,6 +1,5 @@
 package com.record.video.model.remote.response
 
-import com.record.network.model.RecordInfo
 import com.record.video.model.VideoData
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
